공격 상태패턴 구현중
일반공격과 특수공격 구현중

+++ Workpiece/Summoner/Assets/Script/TestSummons/KingSlime.cs
@@ -37,7 +37,7 @@ public class KingSlime : Summon
         base.die();
     }
 
-    public override void takeDamage(int damage)
+    public override void takeDamage(double damage)
     {
         base.takeDamage(damage);
     }
